+++ src/data/chapters.ts
@@ -17,6 +17,7 @@ import { CHAPTER_23 } from './chapter-23';
 import { CHAPTER_24 } from './chapter-24';
 import { CHAPTER_25 } from './chapter-25';
 import { CHAPTER_26 } from './chapter-26';
+import { CHAPTER_27 } from './chapter-27';
 import { CHAPTER_3 } from './chapter-3';
 import { CHAPTER_4 } from './chapter-4';
 import { CHAPTER_5 } from './chapter-5';
@@ -302,4 +303,5 @@ export const CHAPTERS = [
   CHAPTER_24,
   CHAPTER_25,
   CHAPTER_26,
+  CHAPTER_27,
 ] as const;